basic building blocks defined
